Moved from script to js
